Revert "ResourceHandle : Refactor resourceHandle data type to long"
Revert submission 28771115-cherrypicker-L41100030005737508:N83900030090946413

Reason for revert: Droidmonitor created revert due to b/360262704. Will be verifying through ABTD before submission.


Reverted changes: /q/submissionid:28771115-cherrypicker-L41100030005737508:N83900030090946413

+++ tests/tests/tv/src/android/media/tv/tuner/cts/TunerTest.java
@@ -2530,7 +2530,7 @@ public class TunerTest {
     }
 
     private TunerFrontendInfo tunerFrontendInfo(
-            long handle, int frontendType, int exclusiveGroupId) {
+            int handle, int frontendType, int exclusiveGroupId) {
         TunerFrontendInfo info = new TunerFrontendInfo();
         info.handle = handle;
         info.type = frontendType;
@@ -2574,9 +2574,9 @@ public class TunerTest {
         }
     }
 
-    private void assignFeResource(
-            int clientId, int frontendType, boolean expectedResult, long expectedHandle) {
-        long[] feHandle = new long[1];
+    private void assignFeResource(int clientId, int frontendType,
+                                  boolean expectedResult, int expectedHandle) {
+        int[] feHandle = new int[1];
         TunerFrontendRequest request = new TunerFrontendRequest();
         request.clientId = clientId;
         request.frontendType = frontendType;
